Organized the settings somewhat. Added comments to many of them.
Renamed several settings:
    "width" --> "lineWidth"
    "inner" --> "innerRadius"
    "outer" --> "outerRadius"

Added a new setting: "minimumOpacity" -- used to be fixed at 0.18. The behavior (equation used to generate range of opacities) is a little strange, I may add more control over this.

Storing originalTarget (the DOM element) so that we can bring it back later with the new "restore" method.

Fixed ability to stop animating if the target is no longer in the DOM -- broken in a previous commit.

"autoStart" option (defaults to true) now calls "start" method instead of duplicating code.

I was adding some things to the settings object that didn't need to be in there -- because they are calculated based on other settings, and cannot actually be set by passing different values in) -- I removed them, in part in order to achieve better minification. "canvasSize", "canvasSizeCSS", "center" (which was also renamed to "canvasCenter"), & "delay".

I also looked at which settings were repeated 3 or more times, and added them to the var declaration at the top, then added lines like "minimumOpacity = settings.minimumOpacity;" before the first usage -- largely for better minification. I may want to revisit this later and make sure it's not making the code harder to read / think about the cleanest pattern.

Also a few JSLint-related changes.

Plus some general cleanup.