Now generates data URIs from canvas, and uses those to display the animation, instead of using a tall canvas containing all frames ("filmstrip"). This eliminates the "bobbing" that was affecting iOS & some other browsers (using page zoom).
"Filmstrip" approach is still used for browsers lacking the "toDataUrl()" canvas method.

Changed fallback from a PNG "filmstrip" to a JavaScript array of data URIs.

Added an "export" method to generate an array of data URIs.

Now namespacing handlers bound to window, and removing them when the "restore" method is called.

Added more comments to the code.

Other minor changes.